requirements.txt
python-binance==1.0.19
pandas==2.1.4
numpy==1.24.3
ta==0.10.2
requests==2.31.0
websocket-client==1.6.4
python-telegram-bot==20.7
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
